\name{eqse-package}
\alias{eqse-package}
\alias{eqse}
\docType{package}
\title{
\packageTitle{eqse}
}
\description{This package is used to calcuate the analytic standard error of equipercentile equating under single- or random- groups design. 
\packageDescription{eqse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eqse}
\packageIndices{eqse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{Liuhan Cai
\packageAuthor{eqse}

Maintainer: \packageMaintainer{eqse}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
